+++ napari/layers/base/base.py
@@ -6,9 +6,10 @@ from typing import List, Optional
 import numpy as np
 
 from ...components import Dims
+from ...utils.dask import configure_dask
 from ...utils.event import EmitterGroup, Event
 from ...utils.key_bindings import KeymapProvider
-from ...utils.misc import ROOT_DIR, configure_dask
+from ...utils.misc import ROOT_DIR
 from ...utils.naming import magic_name
 from ...utils.status_messages import format_float, status_format
 from ..transforms import ScaleTranslate, TransformChain
